refactor: WeatherInfoRepositoryをprotocolに変更
