+++ ios-training/SceneDelegate.swift
@@ -15,7 +15,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         window = UIWindow(frame: windowScene.coordinateSpace.bounds)
         window?.windowScene = windowScene
         window?.rootViewController = WeatherViewController(
-            weatherInfoRepository: WeatherInfoRepository()
+            weatherInfoRepository: WeatherInfoRepositoryImpl()
         )
         window?.makeKeyAndVisible()
     }
